feat: algumas funções de menu e file reader
